Move caseActivities from components to state
